requirements.txt

# ===== Web 框架 =====
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic-settings>=2.0.0

# ===== 数据库 =====
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pgvector>=0.2.5

# ===== AI/LLM =====
langchain>=0.3.7
langchain-openai>=0.2.5
langchain-community>=0.3.5
langgraph>=0.2.28
langgraph-cli>=0.1.6
langgraph-checkpoint>=3.0.1
langgraph-checkpoint-postgres>=3.0.0
langchain-core>=0.3.15
openai>=1.54.0
python-dotenv>=1.0.0

# ===== 数据处理 =====
pandas>=2.1.0
numpy>=1.26.0
scipy>=1.11.0

# ===== 机器学习 =====
scikit-learn>=1.3.0
statsmodels>=0.14.0
mlxtend>=0.23.0

# ===== 数据可视化 =====
pyecharts>=2.0.0
seaborn>=0.13.0
kaleido>=0.2.0

# ===== 用户认证 =====
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# ===== HTTP请求 =====
requests>=2.31.0

# ===== PDF 处理 =====
pypdf>=3.0.0
pdfplumber>=0.10.0

# ===== 报告生成 =====
reportlab>=4.0.0

# ===== 数据导入导出 =====
openpyxl>=3.1.0
xlsxwriter>=3.1.0
